requirements.txt
streamlit==1.32.0
transformers==4.38.2
torch==2.2.1
Pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.4
plotly==5.19.0
huggingface-hub==0.21.4
tokenizers==0.15.2
